refactor: update import from constants
