--- apps/server/src/constants/cache.ts
@@ -1,3 +1,0 @@
-export const FI_STORE_KEY = "fi";
-export const FIAGRO_STORE_KEY = "fiagro";
-export const CACHE_DURATION = 60 * 30; // 30 minutes
