--- apps/server/src/constants/queue.ts
@@ -1,1 +1,0 @@
-export const NOTION_DATA_SOURCE_FI_CHANNEL = "notion.data-source.get.fi";
